services/paddleocr-service/requirements.txt
# PaddleOCR Layout Detection Service Dependencies

# Deep Learning Framework
paddlepaddle==2.6.0  # CPU version
# paddlepaddle-gpu==2.6.0  # GPU version (uncomment if using GPU)

# OCR Library
paddleocr>=2.7.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging (optional, for JSON logging)
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For testing FastAPI endpoints

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
